apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.9.0
python-multipart>=0.0.12
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
